officialSearchContract -> contractsModeRouting.ts
